requirements.txt
crewai==0.148.0
google-generativeai==0.8.5
python-dotenv==1.1.1
requests==2.32.4
beautifulsoup4==4.13.4
textstat==0.7.8
gradio==5.38.0
langchain-google-genai==2.1.9 